Change doc ID type from Integer to Long for getDocumentLength method.  Consistent with change to doc ID type for getDocumentName method.

+++ core/src/main/java/org/lemurproject/galago/core/retrieval/Retrieval.java
@@ -200,7 +200,7 @@ public interface Retrieval extends Closeable {
    * @return document length
    * @throws IOException
    */
-  Integer getDocumentLength(Integer docid) throws IOException;
+  Integer getDocumentLength(Long docid) throws IOException;
 
   /**
    * Returns the length of a particular document. Where docname
